Source/server/requirements.txt
aes==1.0.0
appdirs==1.4.3
astroid==2.4.1
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
config==0.5.0.post0
cycler==0.10.0
distlib==0.3.0
dnspython==1.16.0
enums==0.0.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
fuzzywuzzy==0.18.0
google==2.0.3
google-api-core==1.17.0
google-auth==1.15.0
google-cloud==0.34.0
google-cloud-speech==1.3.2
googleapis-common-protos==1.51.0
grpcio==1.29.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mongoengine==0.20.0
motor==2.1.0
mysql-connector-python==8.0.20
Naked==0.1.31
numpy==1.18.4
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pylint==2.5.2
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-speech-features==0.6
pytz==2020.1
PyYAML==5.3.1
randomword==1.0.2
RandomWords==0.2.1
requests==2.23.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
sha256==0.1
Shapely==1.7.0
shellescape==3.8.1
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
speech==0.5.2
SpeechRecognition==3.8.1
toml==0.10.0
urllib3==1.25.9
virtualenv==20.0.20
watson-developer-cloud==2.10.1
websocket-client==0.48.0
Werkzeug==1.0.1
wrapt==1.12.1
